backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==25.4.0
bcrypt==5.0.0
bidict==0.23.1
boto3==1.42.44
botocore==1.42.44
CacheControl==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
deepgram-sdk==5.3.2
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
elevenlabs==2.34.0
email-validator==2.3.0
fastapi==0.128.5
firebase_admin==7.1.0
frozenlist==1.8.0
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
jmespath==1.1.0
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
msgpack==1.1.2
multidict==6.7.1
numpy==2.4.2
openai==2.17.0
packaging==26.0
passlib==1.7.4
pgvector==0.4.2
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
PyNaCl==1.6.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-http-client==3.3.7
python-multipart==0.0.22
python-socketio==5.16.1
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rsa==4.9.1
s3transfer==0.16.0
sendgrid==6.12.5
simple-websocket==1.1.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.52.1
structlog==25.5.0
telnyx==4.20.0
tqdm==4.67.3
twilio==9.10.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.5
wrapt==2.1.1
wsproto==1.3.2
yarl==1.22.0
